Stop unconditionally applying hermetic mode
When originally authored, it was not necessary to unconditionally apply
hermetic mode, but I chose to apply it in eager mode to help catch bugs.
Well, multipy is kind of dead, and hermetic mode is causing real
implementation problems for people who want to do fancy Python stuff
from the dispatcher.  So let's yank this mode for now.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: d0bc59ecd0de25cd6bccd782832bd46246e50f10
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116996